Fixes #6. Slightly refactored tests.
