+++ src/vfb_connect/neo/test/neo_tools_tests.py
@@ -4,22 +4,42 @@ Created on Oct 24, 2016
 @author: davidos
 '''
 import unittest
-from ..neo4j_tools import Neo4jConnect, dict_cursor
-import pandas as pd
+from ...cross_server_tools import VfbConnect
 
-class Test_commit(unittest.TestCase):
-    
+# TODO- pull over core nc tests from VFB_neo4j
+
+class NewQueryWrapperTest(unittest.TestCase):
+
     def setUp(self):
-        self.nc = Neo4jConnect('http://localhost:7474', 'neo4j', 'neo4j')
+        # Using this as a convenient way to wrap default neo server call
+        self.vc = VfbConnect()
 
-    def test_neo_create_node(self): 
-        # Quite a minimal test...
-        self.assertTrue(expr=self.nc.commit_list(["CREATE (t:Test { fu: 'bar' })"]),
-                        msg="fu")
 
-    def tearDown(self):
-        self.nc.commit_list(["MATCH (t:Test { fu: 'bar' }) DELETE t"])
+    def test_get_term_info(self):
+        self.assertTrue(self.vc.neo_query_wrapper.get_type_TermInfo(['FBbt_00003686']))
+        self.assertTrue(
+            self.vc.neo_query_wrapper.get_DataSet_TermInfo(['Ito2013']))
+        self.assertTrue(
+            self.vc.neo_query_wrapper.get_anatomical_individual_TermInfo(['VFB_00010001']))
+        self.assertTrue(
+            len(self.vc.neo_query_wrapper.get_TermInfo(['FBbt_00003686', 'VFB_00010001', 'Ito2013'])) == 3)
 
+    def test_get_by_xref(self):
+        self.assertTrue(self.vc.neo_query_wrapper.get_terms_by_xref(['Trh-F-500041'], db='FlyCircuit'))
+
+    def test_get_vfb_id_by_xref(self):
+        self.assertTrue(self.vc.neo_query_wrapper.xref_2_vfb_id(['Trh-F-500041'], db='FlyCircuit'))
+
+    def test_get_xref_by_vfbid(self):
+        self.assertTrue(self.vc.neo_query_wrapper.vfb_id_2_xrefs(['VFB_00014110'], db='FlyCircuit'))
+
+    def test_get_image_by_filename(self):
+        self.assertTrue(
+            self.vc.neo_query_wrapper.get_images_by_filename(
+                'JFRC2_MBON-b1-a.nrrd'))
+        self.assertTrue(
+            self.vc.neo_query_wrapper.get_images_by_filename(
+                'JFRC2_MBON-b1-a.nrrd', dataset='Aso2014'))
 
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']
